Clean. Extract EnumClass from UserType
