+++ ulox/ulox.core/Package/Runtime/Compiler/Compilettes/TypeCompilette.cs
@@ -45,7 +45,6 @@ namespace ULox
         public static TypeCompilette CreateEnumCompilette()
         {
             var compilette = new TypeCompilette();
-            //compilette.AddInnerDeclarationCompilette(new TypeMixinCompilette());
             compilette.AddInnerDeclarationCompilette(new TypeEnumValueCompilette());
             compilette.GenerateCompiletteByStageArray();
             compilette.UserType = UserType.Enum;
